Test/13 board test

+++ src/main/java/noodlezip/community/service/BoardServiceImpl.java
@@ -49,6 +49,7 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public Map<String, Object> findBoardListByCategory(String category, Pageable pageable) {
         Page<BoardRespDto> boardPage = boardRepository.findBoardWithPaginationAndCommunityType(category, pageable);
 
@@ -62,7 +63,11 @@ public class BoardServiceImpl implements BoardService {
     @Override
     @Transactional(readOnly = true)
     public BoardRespDto findBoardById(Long id, Long userId) {
-        return boardRepository.findBoardByBoardIdWithUser(id, userId);
+        BoardRespDto boardRespDto = boardRepository.findBoardByBoardIdWithUser(id, userId);
+        if (boardRespDto == null) {
+            throw new CustomException(ErrorStatus._DATA_NOT_FOUND);
+        }
+        return boardRespDto;
     }
 
     @Override
@@ -88,8 +93,13 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
-    public void deleteBoard(Long boardId) {
+    public void deleteBoard(Long boardId, Long userId) {
         Board board = boardRepository.findById(boardId).orElseThrow(() -> new CustomException(ErrorStatus._DATA_NOT_FOUND));
+
+        if(!board.getUser().getId().equals(userId)) {
+            throw new CustomException(ErrorStatus._FORBIDDEN);
+        }
+
         boardRepository.delete(board);
     }
 }
